requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.21
sqlmodel>=0.0.8
asyncpg>=0.27
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
sqlalchemy>=1.4

fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=1.10.0
pydantic-settings>=0.5
langchain>=0.0.300
chromadb>=0.3.28
faiss-cpu>=1.7.4
openai>=0.27.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
email-validator>=1.3
greenlet>=2.0
aiosqlite>=0.19
